Backoffice atualizado

+++ core/src/main/java/com/example/core/repositories/ClienteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
-    // Se precisar de queries customizadas, declare aqui.
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+    // Se precisares de queries customizadas, podes declarar aqui.
 }
