+++ core/src/main/java/com/example/core/repositories/FuncionariosFixosRepository.java
@@ -5,7 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FuncionariosFixosRepository extends JpaRepository<Funcionariosfixo, Long> {
-    // Aqui você pode adicionar métodos customizados, se necessário.
+public interface FuncionariosFixosRepository extends JpaRepository<Funcionariosfixo, Integer> {
 }
-
