+++ core/src/main/java/com/example/core/repositories/ReciboRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReciboRepository extends JpaRepository<Recibo, Long> {
+public interface ReciboRepository extends JpaRepository<Recibo, Integer> {
     // Se precisar de queries customizadas, declare aqui.
 }
