+++ core/src/main/java/com/example/core/repositories/TipoServicoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoServicoRepository extends JpaRepository<Tiposervico, Long> {
+public interface TipoServicoRepository extends JpaRepository<Tiposervico, Integer> {
 }
